Upgrade to opencv 3.1, add armhf support
Upgrading to opencv 3.1 and javacpp 1.2 adds out-of-the-box support
for armhf.  On raspbian, one must pass -Dsun.arch.abi=gnueabihf in
order for javacpp to load the correct natives.  I have disabled
the frc natives until someone rebuilds them with the frc toolchain.

Changing to opencv 3.1 required some type casts, I'm not sure if
what I've done is best since I haven't touched the internals of
GRIP before.

+++ core/src/main/java/edu/wpi/grip/core/operations/composite/PublishVideoOperation.java
@@ -105,8 +105,8 @@ public class PublishVideoOperation implements Operation {
                             imageLock.wait();
 
                             // Copy the image data into a pre-allocated buffer, growing it if necessary
-                            bufferSize = imagePointer.limit();
-                            if (bufferSize > buffer.length) buffer = new byte[imagePointer.limit()];
+                            bufferSize = (int) imagePointer.limit();
+                            if (bufferSize > buffer.length) buffer = new byte[(int) imagePointer.limit()];
                             imagePointer.get(buffer, 0, bufferSize);
                         }
